requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
certifi==2025.6.15
cfgv==3.4.0
click==8.2.1
distlib==0.4.0
fastapi==0.115.13
filelock==3.18.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httptools==0.6.4
identify==2.6.12
idna==3.10
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.1
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
ruff==0.12.4
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
